Remove parsing related code from collector
